requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
env==0.1.0
filelock==3.0.12
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pipenv==2020.6.2
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.24
virtualenv-clone==0.5.4
whitenoise==5.1.0
wrapt==1.12.1
